Create Assignment3T2ShortestPath.py
I can comprehend the steps at this point, but I'm not getting around at which point in the code the assertion error is being thrown for the short_test. All I know is that my attempt is fundamentally broken. But I got a chance to try list comprehension in my code, so at least there's that :)