Update Assignment3T2ShortestPath.py
I print a LOT to the console throughout this algorithm, but I find that it is helpful for visualizing where mistakes arise. That being said, I can pass all the easy tests, but the medium tests throw assertion errors still. I'm not sure where my algorithm is going wrong there, but the returned results are incorrect. Either way, I'm happy that some of it works :D