Attempted solution to shortest path w/ greedy algo
This algorithm is GREEDY, so it will take the first minimal path it finds locally. Globally, this algorithm will likely not be optimal since the edge weight from one node to another compared to that of the other's neighbors do not always indicate which child node will be used in an OVERALL lesser cost from start to end. The she shortest path from one node to the next may seem like a good option, but the path may redirect you to a series of other nodes, which adds additional weights before you reach the end node. This means that it may or may NOT result in the shortest path.